feat(chunk_fwd_o): overlap A5 stage1 and stage2 heads
Add PR404-style C:V=1:2 per-head signaling, owner-directed BF16 Fixpipe outputs, and two-slot in-flight acknowledgements so Stage1 and Stage2 overlap safely. Validate G=1 at T=64/128/256 and both owner ping-pong slots.

+++ fla/ops/ascendc/gdn/chunk_gdn_fwd/chunk_fwd_o/op_kernel/chunk_fwd_o_a5_constants.h
@@ -27,7 +27,7 @@ constexpr uint32_t CHUNK_FWD_O_DBG_GATE_A_BYTES =
 constexpr uint32_t CHUNK_FWD_O_DBG_ARAW_BYTES =
     static_cast<uint32_t>(CHUNK_FWD_O_A5_BT * CHUNK_FWD_O_A5_BT * 2U); // bf16
 constexpr uint32_t CHUNK_FWD_O_DBG_OSRAW_BYTES =
-    static_cast<uint32_t>(CHUNK_FWD_O_A5_BT * CHUNK_FWD_O_A5_V * sizeof(float)); // fp32
+    static_cast<uint32_t>(CHUNK_FWD_O_A5_BT * CHUNK_FWD_O_A5_V * sizeof(uint16_t)); // bf16
 constexpr uint32_t CHUNK_FWD_O_DBG_APRIME_BYTES =
     static_cast<uint32_t>(CHUNK_FWD_O_A5_BT * CHUNK_FWD_O_A5_BT * sizeof(uint16_t)); // bf16
 constexpr uint32_t CHUNK_FWD_O_DBG_OSPRIME_BYTES =
